Some minor cleanup.
